api/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-keyvault-secrets>=4.8.0
openai>=1.30.0
applicationinsights>=0.11.10
opencensus-ext-azure>=1.1.13
